Use the new console reporter
